pyproject.toml
[project]
name = "pw-mcp-server-http"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "google-genai>=1.16.1",
    "mcp>=1.9.1",
    "playwright>=1.52.0",
]
